+++ CHANGELOG.md
@@ -14,7 +14,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added cli support for setting environment variables in mutators and handlers.
 - Added gRPC extension service definition.
 - Support for nightlies in build/packaging tooling.
-- Added extension regsitry to the store.
+- Added extension registry to the store.
+- Added extension registry support to apid.
 
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024